refactor(product): 음식 디렉토리 분리
