+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,5 +1,7 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.product.domain.Product;
+
 import javax.persistence.*;
 import java.util.UUID;
