+++ src/main/java/kitchenpos/product/domain/JpaProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.product.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
